requirements.txt
pandas==1.0.1
numpy==1.18.1
netCDF4==1.5.3
PyYAML==5.3
matplotlib==3.1.3
scipy==1.5.4
dask==2022.2.0
xarray==0.19.0
envass==0.0.7
dolfyn==0.12.3